Release 1.129.0

+++ classes/entities/userlogins/migration.php
@@ -43,7 +43,7 @@ class migration extends \local_intellidata\entities\migration {
      * @return array
      */
     public function get_sql($count = false, $condition = null, $conditionparams = [], $timestart = null) {
-        $release4 = (float)ParamsHelper::get_release() >= 4.0;
+        $release4 = ParamsHelper::get_release() >= 4.0;
         $sqlwhere = ''; $sqlparams = [];
 
         if ($timestart > 0) {
